refactor: Update Students class to use instance attribute for programming language
This commit updates the Students class in the Classes&Objects.py file to use an instance attribute for the programming language instead of a class attribute. This allows each instance of the Students class to have its own programming language value. The commit also includes a comment explaining that if no value is assigned to the instance attribute, it will take the value of the class attribute by default.